refactor(console): move Button into its own directory
Follows the dropdown/sheet/sidebar convention of one folder per
component. No API changes — consumers still `import { Button } from
'.../components/button'`.
